made the react app and created the nav bar and started working on the create page

+++ backend/controllers/product.controller.js
@@ -50,6 +50,9 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res)  =>{
     const {id} = req.params;
+    if(!isValidObjectId(id)){
+        return res.status(404).json({success: false, message: 'Invalid Product ID'})
+    }
     
     try {
         await Product.findByIdAndDelete(id);
